refactor client-db; include file-access-log file
